automation/requirements.txt
browser-use==0.1.40
python-dotenv==1.0.1
langchain-openai==0.3.0
playwright==1.49.1
aiohttp==3.11.11
aiofiles==24.1.0
pydub==0.25.1
moviepy==2.1.1
websockets==14.1
watchdog==6.0.0
